Moved routes out of server file
Later we can separate related routes into their own files.
