+++ generateData.js
@@ -28,7 +28,7 @@ DB.open(function(err, db) {
             userObjs.push(UserModel.create("Guybrush", "Threepwood", "Guybrush Threepwood", "test1", "test1@test.com", "Threepy", "Hippy Lumberjack"));
             userObjs.push(UserModel.create("Rick", "Sanchez", "Rick Sanchez", "test2", "test2@test.com", "Dirty", "Street Pharmacist"));
             userObjs.push(UserModel.create("Dick", "Dickerson", "Dick Dickerson", "test3", "test3@test.com", "The Fracker", "Money Bags Oil Man"));
-            var insertedIds;
+            var insertedUserIds;
             db.collection('users').insert(userObjs,
                 function(err, results) {
                     if (err) {
@@ -37,7 +37,7 @@ DB.open(function(err, db) {
                     }
                     else {
                         console.log(chalk.bgGreen('Users created in the users collection.'));
-                        insertedIds = results.insertedIds;
+                        insertedUserIds = results.insertedIds;
                     }
                 }
             );
@@ -51,6 +51,7 @@ DB.open(function(err, db) {
                     eventObjs.push(EventModel.create("In Progress Event 1", "USMAY", now.toISOString(), getFollowingWeek(now).toISOString()));
                     eventObjs.push(EventModel.create("In Progress Event 2", "USMKE", now.toISOString(), getFollowingWeek(now).toISOString()));
                     eventObjs.push(EventModel.create("Completed Event 1", "USTWB", getPreviousWeek(getPreviousWeek(now)).toISOString(), getPreviousWeek(now).toISOString()));
+                    var insertedEventIds;
                     db.collection('events').insert(eventObjs,
                         function(err, results) {
                             if (err) {
@@ -59,6 +60,7 @@ DB.open(function(err, db) {
                             }
                             else {
                                 console.log(chalk.bgGreen('Events created in the events collection.'))
+                                insertedEventIds = results.insertedIds;
                             }
                         }
                     );
@@ -73,9 +75,9 @@ DB.open(function(err, db) {
                               }
                               else {
                                   var ideaObjs = [];
-                                  ideaObjs.push(IdeaModel.create("Guybrush's Test Idea", "This is an idea description.", insertedIds[0], [], [], []));
-                                  ideaObjs.push(IdeaModel.create("Rick's Test Idea", "This is Mr. Sanchez\'s brilliant idea.", insertedIds[1], [], [], []));
-                                  ideaObjs.push(IdeaModel.create("Dick's Test Idea", "This is \"The Fracker\'s\" master plan.", insertedIds[2], [], [], []));
+                                  ideaObjs.push(IdeaModel.create("Guybrush's Test Idea", "This is an idea description.", insertedUserIds[0], insertedEventIds[0], [], []));
+                                  ideaObjs.push(IdeaModel.create("Rick's Test Idea", "This is Mr. Sanchez\'s brilliant idea.", insertedUserIds[1], insertedEventIds[1], [], []));
+                                  ideaObjs.push(IdeaModel.create("Dick's Test Idea", "This is \"The Fracker\'s\" master plan.", insertedUserIds[2], insertedEventIds[2], [], []));
                                   db.collection('ideas').insert(ideaObjs,
                                       function(err, results) {
                                           if (err) {
